wifitomqtt: propagate flags even when they haven't changed
This commit fixes the scenario for creating a new network.
Some flags may have been set already before the network was created
in wpa_suppicant. The test for flags changes later on may not indicate
a change, while it's still required to propagate the initial flags.
Changing the related BSS do test the flags changes anyway, so this
commit's impact is only that newly created networks get their flags
propagated.